♻️ Move libs into sub folders of domains
