+++ dashboard/libs/content/.babelrc
