+++ dashboard/libs/content/package.json
