+++ dashboard/libs/content/project.json
@@ -1,7 +1,7 @@
 {
   "name": "dashboard-content",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "dashboard/dashboard-content/src",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "dashboard/libs/content/src",
   "projectType": "library",
   "tags": [],
   "// targets": "to see all targets run: nx show project dashboard-content --web",
