+++ project-manager/libs/content/.babelrc
