+++ project-manager/libs/content/project.json
@@ -1,7 +1,7 @@
 {
   "name": "project-manager-content",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "project-manager/project-manager-content/src",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "project-manager/libs/content/src",
   "projectType": "library",
   "tags": [],
   "// targets": "to see all targets run: nx show project project-manager-content --web",
