+++ project-manager/libs/content/src/lib/components/projectCard/projectCard.stories.tsx
@@ -27,8 +27,8 @@ export const Primary: Story = {
           icon: 'person',
           label: 'Me',
           href: 'https://github.com/anthony-jhoiro',
-        }
+        },
       ],
-    }
+    },
   },
 };
