+++ project-manager/libs/content/src/lib/models/project.ts
@@ -4,9 +4,9 @@ export type ProjectLink = {
   icon: ProjectIconType;
   label: string;
   href: string;
-}
+};
 
 export type Project = {
   name: string;
   links: ProjectLink[];
-}
+};
