+++ project-manager/libs/content/vite.config.ts
@@ -4,7 +4,8 @@ import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
 
 export default defineConfig({
   root: __dirname,
-  cacheDir: '../../node_modules/.vite/project-manager/project-manager-content',
+  cacheDir:
+    '../../../node_modules/.vite/project-manager/project-manager-content',
 
   plugins: [react(), nxViteTsPaths()],
 
@@ -16,13 +17,13 @@ export default defineConfig({
   test: {
     setupFiles: ['./src/test-setup.ts'],
     globals: true,
-    cache: { dir: '../../node_modules/.vitest' },
+    cache: { dir: '../../../node_modules/.vitest' },
     environment: 'jsdom',
     include: ['src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     reporters: ['default'],
     coverage: {
       reportsDirectory:
-        '../../coverage/project-manager/project-manager-content',
+        '../../../coverage/project-manager/project-manager-content',
       provider: 'v8',
     },
   },
