+++ shared/ui/react-storybook/project.json
@@ -4,7 +4,10 @@
   "sourceRoot": "shared/ui/react-storybook/src",
   "projectType": "library",
   "tags": [],
+  "implicitDependencies": [
+    "react-components",
+    "project-manager-content"
+  ],
   "// targets": "to see all targets run: nx show project react-storybook --web",
-  "targets": {},
-  "implicitDependencies": ["react-components", "dashboard-content", "project-manager-content"]
+  "targets": {}
 }
